+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
 ## [Unreleased]
 
-## [0.1.5] - 2025-01-14
+## [0.1.6] - 2025-01-14
 
 ### Added
 
@@ -33,6 +33,15 @@
 
 - Button 컴포넌트의 `type` prop이 올바르게 전달되도록 수정
   - `type="submit"` 등 명시적 type prop이 정상 작동
+
+## [0.1.5] - 2025-12-30
+
+### Added
+
+- Typography 토큰을 CSS 변수로 노출
+  - 각 typography 스타일(h1-h6, subtitle-1/2, body-1/2, caption, overline)의 fontSize, fontWeight, lineHeight, letterSpacing을 CSS 변수로 노출
+  - 사용처에서 Tailwind CSS를 통해 `text-caption`, `text-h1` 등의 클래스를 바로 사용 가능
+  - CSS 변수를 통해 런타임에 typography 스타일 커스터마이징 가능
 
 ## [0.1.4] - 2025-12-30